requirements.txt
streamlit>=1.55.0
anthropic>=0.40.0
google-cloud-aiplatform
google-auth-oauthlib
